+++ packages/headless/src/app/commerce-engine/commerce-engine.ssr.ts
@@ -87,6 +87,7 @@ export interface CommerceEngineDefinition<
   > {}
 
 /**
+ * @internal
  * Initializes a Commerce engine definition in SSR with given controllers definitions and commerce engine config.
  * @param options - The commerce engine definition
  * @returns Three utility functions to fetch the initial state of the engine in SSR, hydrate the state in CSR,
@@ -97,10 +98,10 @@ export function defineCommerceEngine<
     CommerceEngine,
     Controller
   >,
->({
-  controllers: controllerDefinitions,
-  ...engineOptions
-}: CommerceEngineDefinitionOptions<TControllerDefinitions>): CommerceEngineDefinition<TControllerDefinitions> {
+>(
+  options: CommerceEngineDefinitionOptions<TControllerDefinitions>
+): CommerceEngineDefinition<TControllerDefinitions> {
+  const {controllers: controllerDefinitions, ...engineOptions} = options;
   type Definition = CommerceEngineDefinition<TControllerDefinitions>;
   type BuildFunction = Definition['build'];
   type FetchStaticStateFunction = Definition['fetchStaticState'];